+++ pcweb/pages/docs/__init__.py
@@ -109,6 +109,7 @@ outblocks = []
 
 
 def get_component(doc: str, title: str):
+    return
     if doc.endswith("-style.md"):
         return
