+++ pcweb/pages/landing_page_components/landing_page_style.py
@@ -1,0 +1,47 @@
+import reflex as rx
+
+feature_button_style = dict(
+    color="#A19DD5",
+    font_size="sm",
+    text_decoration="none",
+    margin_right="2em",
+)
+
+heading_1_style = dict(
+    font_size="54px",
+    background_image="linear-gradient(95deg, #D6D6ED 42.14%, #727280 63.21%)",
+    text_align="left",
+    background_clip="text",
+    font_weight="bold",
+    line_height="1",
+)
+
+heading_2_style = dict(
+    font_size="54px",
+    max_width="650px",
+    color="#6C6C81",
+    text_align="left",
+    font_weight="bold",
+    line_height="1",
+)
+
+get_started_button_style = dict(
+    border_radius="8px",
+    padding_x="2em",
+    width="325px",
+    height="80px",
+    font_size="30px",
+    background="linear-gradient(180deg, #6151F3 0%, #5646ED 100%)",
+    box_shadow="0px 2px 9px -4px rgba(64, 51, 192, 0.70), 0px 0px 6px 2px rgba(255, 255, 255, 0.12) inset, 0px 0px 0px 1px rgba(255, 255, 255, 0.09) inset",
+    _hover={},
+)
+
+get_demo_button_style = dict(
+    font_size="30px",
+    width="325px",
+    height="80px",
+    border_radius="8px",
+    border="2px solid rgba(186, 199, 247, 0.12)",
+    background="rgba(161, 157, 213, 0.03)",
+    backdrop_filter="blur(2px)",
+)
