+++ pcweb/templates/webpage.py
@@ -103,7 +103,7 @@ def webpage(path: str, title: str = DEFAULT_TITLE, props=None) -> Callable:
             # Import here to avoid circular imports.
             from pcweb.components_webpage.footer import footer
             from pcweb.components_webpage.navbar import navbar
-            from pcweb.components_webpage.sidebar import sb
+            from pcweb.components.sidebar import sb
 
             # Wrap the component in the template.
             return rx.flex(
